week09/tests/requirements.txt
pytest==7.4.0
psycopg2-binary==2.9.9
pymongo==4.5.0
pytest-cov==4.1.0
